Add a local learner to each processor, along with training and prediction logic.
LocalLearnerProcessor already does most of these things, but decided to implement
from scrach to better understand what's going on.

Next up: Create a processor that connects the input stream with the first learner,
and the last learner with the first. This will allow us to have a common event
type for all learners (BoostEvent), and provide more encapsulation.