Change the topology and communication pattern, read on for details.
We introduce a new processor, BoostModelProcessor, that coordinates the updates of the models.

It sits between the models and the input, and is responsible for two things:
When new events come in from the input, it attaches the current boosting model to the event,
and sends it on to the first learner who then propagates it down the pipeline.
The model should get updated as it moves down the pipeline.
For that we created a new event, BoostContentEvent that contains an InstanceContentEvent,
and a BoostingModel.

When the BoostModelProcessor  receives a BoostContentEvent from the last learner, that means that the instance
has made its way round the pipeline, so a prediction can be output, and a new instance
fed into it.
Note that currently the output is unused, and the processor does not wait before the previous
input has made its round trip before sending the next.

The BoostingModel is an interface (very much unfinished) that should serve as a basis for all
boosting algorithms. It is supposed to facilitate the per-learner update of the boosting model,
and maintain its state. ToyBoost is a toy example of such a model, OzaBoost will follow in the next commit.
Our assumption here is that it is possible to update the boosting model, using only knowledge
about the "current" weak learner and the training instance.

Next up we need to handle prediction, figure out if it's necessary to block sending new instances
before the previous one has returned (seems likely), and how can we make the algorithm have actual
parallelism.