Blocking sync for Boosting, using an in-memory buffer.
In this design we buffer the elements sent in from the source until the preceding
elements have been fully processed and output from the boosting pipeline.

If the last event is received at the input but there are still events in the buffer,
we just flush them down the pipeline. The accumulation of events depends on the
speed of the learning algorithm: If we are able to get through the pipeline fast,
not many events should accumulate in the buffer.