Initial skeleton for POCA.
Did a rough topology, with one input processor, one parallel processor for the weak
leaners and one for the model aggregation/update and output.
Decided to try out key-based connections for going from the model aggregator to the
parallel weak learners, will have to figure out how to accomplish that.

Initial idea is to send out events with keys corresponding to the WL id as soon
as the messages from all the preceding WLs have been sent to the aggregator (for that
iteration/instance).

Up next:

Logic for the processors. Specifically instance handling/learning for the WL processors,
model aggregation and update for the aggregator (ModelProcessor), figureout what state
do the POCA events need to carry.