Some simple debugging output for the skeleton to figure out what happens. Read on for details.
This skeleton prints out information about each stage of the pipeline.
Here's what I was able to figure out when running the same task (MyTask) in local execution
and threads with 1 thread:

It seems like when using local execution, the execution is done in "event" sequences:
The first event passes through the whole pipeline of processors, then the second passes through the whole
pipeline and so on. So we finish the processing of one **event** before moving to the next.

In threads it seems like we iterate over stages/processors instead:
All the events pass through the first processor, then they all pass through the
second processor and so on. So it seems like we finish with one **processor**
before moving to the next.

What this does (when running **threads with 1 thread**) is that an event that is broadcast,
say from the input processor to all the weak learner processors, is shared between those
processors, and since the last weak learner in that queue (WL 3) is the last to modify it,
it retains the classifierID value from that.
This is different to the local execution, because there each event passes through the whole
pipeline in sequence so processor instances never share objects.

Example: We have ensembleSize == 3, and 3 data points. The events enter at the input processor,
and all the events pass trough it. We then move on to the weak learner processors. All events are
bcast to all the WLs. Event 1 will pass through WLs 1,2,3 in sequence, and each modifies the event
that is shared between then.
**After** that is done, only then is the event passed on to the next stage.
Since WL 3 is the last in row to modify the event, each WL sends to the ModelAggregator the event,
as modified by WL 3, meaning that all of them will have classifierID == 3.

It seems like we will need to create copies/new events before passing them on to the stream
in order to guarantee that they are not shared.