First attempt to route events from ModelAggregator back to WLs. Currently wrong, read on for details.
My assumption was that if I explicitly set the key to _x_ the event should be routed the _x_ processor.
Turns out keyed streams don't work that way, because they need to "project down" the key to the available
parallelism. So what actually happens is that the key is used to create a hash that is then modulo
by the parallelism to route it to one of the parallel instances.

For example I had parallelism 3 for the WLs, and I create an event in the ModelProcessor that I would
like to route to WL 0, so I set it's key to 0. What actually happens is that the key (String "0") is
used to create a hash value (in this case 677) which is modulo with the parallelism (677 % 3 == 2)
to get the key, 2, which means that this event gets routed to the WL with processor ID 2, instead
of 0 as I intended.

It seems like to achieve the point-to-point routing that I want, I'll have to use multiple streams,
which connect the ModelProcessor with specific WLs (?). Not sure if that's possible to do when using
one processor with parallelism.

It seems like the parallel instances of the WLs do _not_ share state, so at least that is fine.
But I'm not sure if I can do point-to-point routing when only one processor instance is available.
I might have to "reverse-engineer" the key allocation, s.t. I can manufacture the hash to fit my
intended end result. Or I can fiddle with the engine to make it possible to use explicit keys
for processors.

Another option is to try to hack around a shuffle distribution.
That would mean I would have to block sending later events before sending the following ones, but
I think that's fine because we have to do it anyway, as to get the state in WL x we must already
have the state for WL x-1. So that could work.