Aggregator <-> WeakLearners routing, stage 2.
Decided to go the shuffle way, with a bit of hacky approach.

Ideally I would like to send events from the aggregator to the WLs
in order, but that doesn't seem possible due to the way that
events are routed for shuffling, meaning they get the key
counter % parallelism, where counter takes values in [1, parallelism],
meaning that the WL with processor ID 1 will always get the first event,
and the one with ID 0 with always get the last.

But, since we are doing synchronized model updates anyway, the order in
which we update the boosting model of the individual WLs should not matter.
This is bound to get more complicated if there are events being sent from
the WLs to the aggregator while we still have not finised processing the previous
one, (similar to LocalBoost), but that could also be handled in a manner similar
to what we are doing there, with a buffer and blocking, at the expense of speed
and scalability.