Aggregator <-> WL routing, stage 3. Read on for details.
Previously I was maintaining one array of WL state, meaning that we should only
be affecting the boosting state of one instance/round at any time. When using threads
this breaks down, because different instances can arrive at the WLs and from
there to the aggregator in parallel, so if we only had one WL state array, event
2 might be affecting the state array of event 1 etc.

I've now changed that to use a concurrent map that maps from instance ID to an
array of WL states. That means that even if events from different instances
arrive at the aggregator, they will only be affecting the state that corresponds
to that instance.

This is definitely an improvement on what we had before, but it could also be
wrong: What happens when event 2 passes through the WLs, when the boosting model
updates from event 1 have not happened yet? Then event 2 will be processed with
"stale" WLs and passed on to the model aggregator. Say event 2 passes through WL
2, which is in round 1 still, on to the aggregator. In the meantime, we send the
boosting update resulting from event 1 to the WLs. Now WLs are (?) in round 2,
and event 2 arrives at WL 1. The prediction will now be wrong no? The boosting
model has changed between WL 1 and WL 2, but we have already sent WL 2's errors
to the aggregator. Seems like a blocking mechanism is needed here as well.
Or we just accept that there might be inconsistencies in the model, but how
does that affect the convergence rate and guarantees?

Also, I might be confusing a need to broadcast the boosting model back to WLs,
with the need to send the errors for an instance to the downstream WLs from
a preceding WL.
The "chain" dependency exists purely in the training of the WLs, in order to train
WL x, we need all the errors from WLs [1...x-1]. So not sure if the boosting model
update plays into this, that actually seems like a broadcast, or local process.